Move XADPCM.h file into audio folder
